@@ -1,30 +1,27 @@
 package com.appsbysha.ohboy
 
-import android.app.Activity
 import android.content.Intent
 import android.graphics.Bitmap
+import android.graphics.BitmapFactory
 import android.os.Bundle
 import android.provider.MediaStore
 import android.view.Menu
 import android.view.MenuItem
 import android.view.View
+import android.widget.EditText
 import android.widget.ImageView
 import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.Observer
-import androidx.lifecycle.ViewModelProviders
+import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import com.appsbysha.ohboy.UiUtils.Companion.dateFormatter
 import com.appsbysha.ohboy.adapters.LineAdapter
-import com.appsbysha.ohboy.database.LineViewModel
-import com.appsbysha.ohboy.database.LineViewModelFactory
-import com.appsbysha.ohboy.database.SayingViewModel
-import com.appsbysha.ohboy.database.SayingViewModelFactory
+import com.appsbysha.ohboy.database.*
 import com.appsbysha.ohboy.entities.Line
 import com.appsbysha.ohboy.entities.LineType
 import com.appsbysha.ohboy.entities.Saying
-import com.appsbysha.ohboy.entities.SayingWithLines
-import kotlinx.android.synthetic.main.activity_add_saying.*
 import java.io.ByteArrayOutputStream
 import java.io.FileNotFoundException
 import java.io.IOException
@@ -33,15 +30,18 @@ import java.util.*
 class AddSayingActivity : AppCompatActivity() {
     private lateinit var sayingViewModel: SayingViewModel
     private lateinit var lineViewModel: LineViewModel
+    private lateinit var childViewModel: ChildViewModel
     private var childId = 0
+    private var childPic: ByteArray? = null
     private lateinit var editTextTitle: TextView
+    private lateinit var dd : TextView
+    private lateinit var MM : TextView
+    private lateinit var yyyy  : TextView
     private lateinit var editTextDescription: TextView
-    private var sayingDate: String? = null
     private lateinit var addPhotoFromGallery: ImageView
     private var sayingImage: ImageView? = null
     private var sayingByteImage: ByteArray? = null
     lateinit var saying: Saying
-    private var listOfLines: MutableList<Line> = mutableListOf()
     private lateinit var linesRecyclerView: RecyclerView
     private lateinit var linesAdapter: LineAdapter
 
@@ -50,7 +50,11 @@ class AddSayingActivity : AppCompatActivity() {
         setContentView(R.layout.activity_add_saying)
 
         editTextTitle = findViewById(R.id.edit_text_title)
+        dd   = findViewById<EditText>(R.id.text_view_dd)
+        MM   = findViewById<EditText>(R.id.text_view_MM)
+        yyyy = findViewById<EditText>(R.id.text_view_yyyy)
         addPhotoFromGallery = findViewById(R.id.upload_photo)
+        addPhotoFromGallery.visibility = View.GONE
         addPhotoFromGallery.setOnClickListener(View.OnClickListener {
             startActivityForResult(Intent(Intent.ACTION_PICK,
                     MediaStore.Images.Media.INTERNAL_CONTENT_URI), GET_FROM_GALLERY)
@@ -63,47 +67,66 @@ class AddSayingActivity : AppCompatActivity() {
         title = "Add Saying"
         val intent = intent
         childId = intent.getIntExtra(MainActivity.extra_childId, 0)
+    //    childPic = intent.getByteArrayExtra(MainActivity.extra_childpic)
+
+
         saying = Saying(sayingId = UUID.randomUUID().toString(), childId = childId)
+
 
         linesRecyclerView = findViewById(R.id.linesRecyclerView)
         linesRecyclerView.layoutManager = LinearLayoutManager(this)
         linesRecyclerView.setHasFixedSize(true)
         linesAdapter = LineAdapter()
-        linesAdapter.setLines(listOfLines, true)
+        linesAdapter.isEditable = true
+        //linesAdapter.setLines(listOfLines, true)
         linesRecyclerView.adapter = linesAdapter
 
-        sayingViewModel = ViewModelProviders
-                .of(this, SayingViewModelFactory(this.application, childId))
+        sayingViewModel = ViewModelProvider(this, SayingViewModelFactory(this.application, childId))
                 .get(SayingViewModel::class.java)
 
+
+         sayingViewModel.fetchChildData().observe(this){child->
+            childPic = child.profilePic
+        }
+
         setButtons()
-        lineViewModel = ViewModelProviders.of(this, LineViewModelFactory(this.application, childId, saying))
+        lineViewModel = ViewModelProvider(this, LineViewModelFactory(this.application, childId, saying))
                 .get(LineViewModel::class.java)
 
-        lineViewModel.allLines.observe(this, androidx.lifecycle.Observer { lines -> linesAdapter.setLines(lines, true) }) //was listOfLines
+        lineViewModel.allLines.observe(this, Observer { lines -> linesAdapter.setLines(lines, true) }) //was listOfLines
 
 
     }
 
     private fun setButtons() {
-        saying_child_button.setOnClickListener {
-            val newLine = Line(UUID.randomUUID().toString(), saying.sayingId, 0, "", LineType.CHILD_LINE.value, null)
+        val sayingChildButton = findViewById<ImageView>(R.id.saying_child_button)
+        childPic?.let {
+        val bitmap = BitmapFactory
+            .decodeByteArray(it, 0, it.size)
+        sayingChildButton.setImageBitmap(bitmap)}
+        sayingChildButton.setOnClickListener {
+            val newLine = Line(UUID.randomUUID().toString(), saying.sayingId, 0, "", LineType.CHILD_LINE.value, null, childPic)
             //     lineViewModel.insert(newLine)
-            listOfLines.add(newLine)
-            linesAdapter.setLines(listOfLines, true)
+       //     listOfLines.add(newLine)
+            linesAdapter.addLine(newLine)
+            //linesAdapter.setLines(listOfLines, true)
         }
-        brackets_button.setOnClickListener {
-            val newLine = Line(UUID.randomUUID().toString(), saying.sayingId, 0, "", LineType.BRACKETS.value, null)
+        val bracketsButton = findViewById<TextView>(R.id.brackets_button)
+        bracketsButton.setOnClickListener {
+            val newLine = Line(UUID.randomUUID().toString(), saying.sayingId, 0, "", LineType.BRACKETS.value, null, null)
             //    lineViewModel.insert(newLine)
-            listOfLines.add(newLine)
-            linesAdapter.setLines(listOfLines, true)
-        }
-        other_person_button.setOnClickListener {
-            val newLine = Line(UUID.randomUUID().toString(), saying.sayingId, 0, "", LineType.OTHER_PERSON_LINE.value, "")
-            //    lineViewModel.insert(newLine)
-            listOfLines.add(newLine)
-            linesAdapter.setLines(listOfLines, true)
+        //    listOfLines.add(newLine)
+       //     linesAdapter.setLines(listOfLines, true)
+            linesAdapter.addLine(newLine)
 
+        }
+        val otherPersonButton = findViewById<ImageView>(R.id.other_person_button)
+        otherPersonButton.setOnClickListener {
+            val newLine = Line(UUID.randomUUID().toString(), saying.sayingId, 0, "", LineType.OTHER_PERSON_LINE.value, "", null)
+            //    lineViewModel.insert(newLine)
+        //    listOfLines.add(newLine)
+       //     linesAdapter.setLines(listOfLines, true)
+            linesAdapter.addLine(newLine)
         }
 
 
@@ -113,23 +136,20 @@ class AddSayingActivity : AppCompatActivity() {
 
         val title = editTextTitle.text.toString()
 
+        val sayingDate = dateFormatter(
+            dd.getText().toString(),
+            MM.getText().toString(),
+            yyyy.getText().toString()
+        )
+
+
         saying.title = title
         saying.date = sayingDate
         saying.photo = sayingByteImage
 
         //  val sws: SayingWithLines = SayingWithLines(saying, listOfLines)
-        /*var value = text_view_dd.text.toString()
-        val day: Int = value.toInt()
-        value = text_view_MM.text.toString()
-        val month: Int = value.toInt()
-        value = text_view_yyyy.text.toString()
-        val year: Int = value.toInt()
-        val calendar = Calendar.getInstance()
-        calendar[year, month] = day
-        val format = SimpleDateFormat("dd/MM/yyyy", Locale.getDefault())
-        sayingDate = format.format(calendar.time)*/
 
-        for (line in listOfLines) {
+        for (line in linesAdapter.lines) {
             lineViewModel.insert(line)
         }
         sayingViewModel.insert(saying)
@@ -157,10 +177,10 @@ class AddSayingActivity : AppCompatActivity() {
         super.onActivityResult(requestCode, resultCode, data)
         var bitmap: Bitmap? = null
         //Detects request codes
-        if (resultCode == Activity.RESULT_OK) {
+        if (resultCode == RESULT_OK) {
             when (requestCode) {
                 GET_FROM_GALLERY -> {
-                    val selectedImage = data!!.data
+                    val selectedImage = data?.data
                     try {
                         bitmap = MediaStore.Images.Media.getBitmap(this.contentResolver, selectedImage)
                     } catch (e: FileNotFoundException) { // TODO Auto-generated catch block
@@ -169,7 +189,7 @@ class AddSayingActivity : AppCompatActivity() {
                         e.printStackTrace()
                     }
                 }
-                CAMERA_PIC_REQUEST -> bitmap = data!!.extras["data"] as Bitmap
+                CAMERA_PIC_REQUEST -> bitmap = data?.extras?.get("data") as Bitmap
             }
             if (bitmap != null) {
                 sayingImage?.setImageBitmap(bitmap)
