@@ -1,7 +1,6 @@
 package com.appsbysha.ohboy.entities
 
 import androidx.room.Entity
-import androidx.room.PrimaryKey
 
 @Entity(tableName = "lines_table", primaryKeys = ["lineId", "lineSayingId"])
 data class Line(
@@ -10,8 +9,26 @@ data class Line(
         var position: Int,
         var description: String,
         var type: Int,
-        var other_person: String?
+        var other_person: String?,
+        var childPic: ByteArray?
 ) {
+        override fun equals(other: Any?): Boolean {
+                if (this === other) return true
+                if (javaClass != other?.javaClass) return false
+
+                other as Line
+
+                if (childPic != null) {
+                        if (other.childPic == null) return false
+                        if (!childPic.contentEquals(other.childPic)) return false
+                } else if (other.childPic != null) return false
+
+                return true
+        }
+
+        override fun hashCode(): Int {
+                return childPic?.contentHashCode() ?: 0
+        }
 
 }
 
